finagle-core: Introduce ServerDispatcher and AnnotatedTransport.
Problem:

Clients annotate WireSend and WireRecv at the bottom of the finagle
stack. There is no parallel for a Server. In addition, we would like
to have a simple type for wrapping Transports to be Annotated.

Solution:

Introduce the ServerDispatcher type, which has two methods dispatch
and handle. Introduce the AnnotatedTransporter type as well,
which can wraps any Transporter so that a WireRecv can be
emitted upon receipt and a WireSend upon a successful write.